Move types used across components to a common place
